Add actions to warnings.
This represents more general actions that could be applied to address warnings. The first action kind is an edit, which is like a fix with caveats, and which needs an indication of specific user intent to apply.

This change also adds content-to-slot-declarations as a default rule, using edit actions to delineate changes which are safe from those that aren't.